requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
astroid==2.3.3
audioread==2.1.8
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coverage==5.1
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-heroku==0.3.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
ffmpeg-python==0.2.0
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
isort==4.3.21
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
librosa==0.7.2
llvmlite==0.32.0
Markdown==3.2.1
mccabe==0.6.1
norbert==0.2.1
numba==0.49.0
numpy==1.18.2
opt-einsum==3.2.1
pandas==0.25.1
protobuf==3.11.3
psycopg2==2.8.5
pycparser==2.20
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2019.3
requests==2.23.0
resampy==0.2.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
SoundFile==0.10.3.post1
spleeter==1.5.1
sqlparse==0.3.1
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.11.2
